pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "febot"
version = "0.1.0"
description = "Slack RAG bot for FE (基本情報技術者試験) study support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "slack-bolt>=1.18.0",
  "boto3>=1.39.0",
  "chromadb>=0.5.0",
  "openai>=1.40.0",
  "python-dotenv>=1.0.0",
  "pypdf>=5.0.0",
  "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.0",
  "pytest>=8.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "B", "UP", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = ["error"]
